+++ src/main/java/com/worldcup/BackWorldCup/repository/GameRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.worldcup.BackWorldCup.model.Game;
 
-public interface GameRepository extends JpaRepository<Game, Integer>{
+public interface GameRepository extends JpaRepository<Game, Long>{
 
+	
 }
